Move BN and RSA utility functions to new files
