+++ lib/crypto/c_src/Makefile.in
@@ -73,7 +73,9 @@ LIBDIR = $(PRIVDIR)/lib/$(TARGET)
 
 CRYPTO_OBJS = $(OBJDIR)/crypto$(TYPEMARKER).o \
 	$(OBJDIR)/atoms$(TYPEMARKER).o \
-	$(OBJDIR)/engine$(TYPEMARKER).o
+	$(OBJDIR)/bn$(TYPEMARKER).o \
+	$(OBJDIR)/engine$(TYPEMARKER).o \
+	$(OBJDIR)/rsa$(TYPEMARKER).o
 CALLBACK_OBJS = $(OBJDIR)/crypto_callback$(TYPEMARKER).o
 NIF_MAKEFILE = $(PRIVDIR)/Makefile
 CRYPTO_STATIC_OBJS = $(OBJDIR)/crypto_static$(TYPEMARKER).o\
